fix: make the timer configs int

+++ SuperVikingKart/SuperVikingKart.cs
@@ -21,8 +21,8 @@ namespace SuperVikingKart
 
         public static SuperVikingKart Instance;
         public static CustomLocalization Localization = LocalizationManager.Instance.GetLocalization();
-        public static ConfigEntry<float> CartRespawnTimeConfig;
-        public static ConfigEntry<float> BuffBlockRespawnTimeConfig;
+        public static ConfigEntry<int> CartRespawnTimeConfig;
+        public static ConfigEntry<int> BuffBlockRespawnTimeConfig;
         private static ConfigEntry<bool> _debugConfig;
         private Harmony _harmony;
 
@@ -30,13 +30,13 @@ namespace SuperVikingKart
         {
             Instance = this;
             _debugConfig = Config.Bind("General", "Debug", false, "Enable debug logging");
-            CartRespawnTimeConfig = Config.Bind("General", "CartRespawnTime", 10f,
+            CartRespawnTimeConfig = Config.Bind("General", "CartRespawnTime", 10,
                 new ConfigDescription("Time in seconds before a destroyed cart respawns. Server synced value.",
-                    new AcceptableValueRange<float>(0f, 300f),
+                    new AcceptableValueRange<int>(0, 300),
                     new ConfigurationManagerAttributes { IsAdminOnly = true }));
-            BuffBlockRespawnTimeConfig = Config.Bind("General", "BuffBlockRespawnTime", 10f,
+            BuffBlockRespawnTimeConfig = Config.Bind("General", "BuffBlockRespawnTime", 10,
                 new ConfigDescription("Time in seconds before a collected buff block reappears. Server synced value.",
-                    new AcceptableValueRange<float>(0f, 300f),
+                    new AcceptableValueRange<int>(0, 300),
                     new ConfigurationManagerAttributes { IsAdminOnly = true }));
 
             _harmony = new Harmony(PluginGUID);
